Update paging helpers to use 32 bit addresses

+++ tile/kernel/page.h
@@ -4,6 +4,7 @@
 #include <kernel/mci.h>
 #include <kernel/memory.h>
 #include <kernel/uart.h>
+#include <limits.h>
 #include <stdint.h>
 
 #define VIRT_BITMAP_SIZE (VADDR_SPACE_SIZE / PAGE_SIZE / 32)
@@ -27,7 +28,7 @@ void map_smc();
 uint32_t* virt_page_alloc(int flags);
 int virt_page_free(uint32_t* addr);
 
-void create_mapping(uint32_t v_addr, uint64_t p_addr, uint32_t size, int flags);
+void create_mapping(uint32_t v_addr, uint32_t p_addr, uint32_t size, int flags);
 int addr_is_mapped(uint32_t* addr);
 
 uint32_t* pgd_offset(uint32_t* pgd, uint32_t addr);
@@ -35,15 +36,15 @@ uint32_t* pmd_offset(uint32_t* pmd, uint32_t addr);
 uint32_t* pmd_alloc(uint32_t* pgd, uint32_t addr);
 void pmd_clear(uint32_t* pgd, uint32_t addr);
 void pte_clear(uint32_t* pmd, uint32_t addr);
-void pmd_insert(uint32_t* pmd, uint32_t v_addr, uint64_t p_addr, int flags);
+void pmd_insert(uint32_t* pmd, uint32_t v_addr, uint32_t p_addr, int flags);
 
 int pmd_is_page_table(uint32_t* pmd);
 uint32_t* pmd_to_page_table(uint32_t* pmd);
 uint32_t pmd_section_to_addr(uint32_t pmd);
 uint32_t pte_to_addr(uint32_t pte);
 
-uint32_t create_pmd_section(uint64_t p_addr, int flags);
+uint32_t create_pmd_section(uint32_t p_addr, int flags);
 uint32_t create_pmd_page_table(uint32_t* page_table);
-uint32_t create_pte(uint64_t p_addr, int flags);
+uint32_t create_pte(uint32_t p_addr, int flags);
 
 #endif
